[ADD] request para booklist, friend, review y userlist. store y update también hecho.

+++ app/Http/Controllers/Api/FollowsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\FollowRequest;
 use App\Http\Resources\FollowsCollection;
 use App\Http\Resources\FollowsResource;
 use App\Models\Follow;
@@ -22,18 +23,48 @@ class FollowsController extends Controller
         return new FollowsResource($follow);
     }
 
-    public function delete($id)
+    public function unfollowAuthor($id_user, $id_follow)
     {
         try {
-            $follow = Follow::findOrFail($id);
-            $follow->delete();
-            return response()->json(['message' => 'La entrada de la tabla book_list ha sido eliminada correctamente'], 200);
+            $follow = Follow::where('id_user', $id_user)
+            ->where('id_author', $id_follow)->first();
+            if ($follow->id_publisher != null) {
+                $follow->id_author = null;
+            } else{
+                $follow->delete();
+            }
+            return response()->json(['message' => 'Habeis dejado de seguir a este autor/a'], 200);
         } catch (\Exception $e) {
-            return response()->json(['error' => 'No se pudo eliminar la entrada de la tabla book_list'], 500);
+            return response()->json(['error' => 'No se pudo dejar se seguir correctamente'], 500);
         }
     }
 
-    public function store(Request $request){
+    public function unfollowPublisher($id_user, $id_follow)
+    {
+        try {
+            $follow = Follow::where('id_user', $id_user)
+                ->where('id_publisher', $id_follow)->first();
+            if ($follow->id_author != null) {
+                $follow->id_publisher = null;
+            } else{
+                $follow->delete();
+            }
+            return response()->json(['message' => 'Habeis dejado de seguir a esta editorial'], 200);
+        } catch (\Exception $e) {
+            return response()->json(['error' => 'No se pudo dejar se seguir correctamente'], 500);
+        }
+    }
 
+    public function store(FollowRequest $request){
+        $follow = new Follow();
+        $follow->id_user = $request->get('id_user');
+        if ($request->get('id_author')){
+            $follow->id_author = $request->get('id_author');
+        }
+        if ($request->get('id_publisher')){
+            $follow->id_publisher = $request->get('id_publisher');
+        }
+        $follow->save();
+        return new FollowsResource($follow);
     }
 }
